split script into library and CLI portions
