+++ .github/workflows/proposals-order.yml
@@ -19,6 +19,6 @@ jobs:
           fetch-depth: 0
       - uses: ljharb/actions/node/install@main
         name: 'nvm install lts/* && npm install'
-      - run: node ./scripts/check-proposals-order.mjs --base "$BASE_SHA"
+      - run: npm run check-proposals-order -- --base "$BASE_SHA"
         env:
           BASE_SHA: ${{ github.event.pull_request.base.sha }}
