+++ package.json
@@ -3,7 +3,7 @@
   "scripts": {
     "a2s": "node ./scripts/agenda-to-schedule.mjs",
     "build": "./scripts/generate.js",
-    "check-proposals-order": "node ./scripts/check-proposals-order.mjs",
+    "check-proposals-order": "node ./scripts/check-proposals-order.cli.mjs",
     "lint": "markdownlint-cli2",
     "lint:fix": "markdownlint-cli2 --fix",
     "test": "npm run lint && npm run test:unit",
